fix(E.1.2.C leg 0): the fan-shift defect — exit-bandit learning reachable for the first time since v5.14.1.F
EventLoop_DrainPostFill forwarded 10 positional args into OneCore's 12-param
signature after f973b5c inserted confidence_ic_variant mid-signature: the
exit-bandit enable flag bound the IC-variant slot and the 0.001 fee literal
truncated into the int enable (=0, always). The reward UPDATE was dead on
live AND backtest; flipping the flag in live would have driven an out-of-
range IC variant -> constant-0 IC -> spurious drift auto-kill.

Closure is structural, not positional: both signatures FULLY de-defaulted
(arity mismatch = compile error at every caller; the compiler enumerated
exactly the 33 test sites the a-class pre-measured), and the per-node facts
are no longer parameters at all -- exit_bandit_enabled + the counterfactual
taker fee derive from the threaded PerNodeCfg slice at point of use. This
COMPLETES the recorded WIP2d-1.B.1 fee-param deletion; SlowPath.hpp's
"param chain DELETED" comment is finally true.

ONE shared binder: tt::EngineCommon_DrainPostFill (live drainer + both
backtest driver sites; EngineSharded_SlowPath_DrainPostFill deleted). The
backtest now forwards drift knobs + ic_variant + per-node cfg -- live
parity by construction; byte-inert at default cfg (drift floor 0).

A-2: confidence_ic_variant registry max 4->0 (WARN_ON_CLAMP surfaces any
clamp at parse; 0 is the only registered variant) + a belt clamp in OneCore.
Backtest completion now saves all four bandit-state files (exit + both
Thompson pools were live-shutdown-only; backtest learning evaporated).

Class guard armed: -Werror=float-conversion on engine + controller_test
(PARTIAL guard -- type-crossing re-maps only; de-defaulting is the full
closure). It immediately caught 4 stale test callers passing 0.10 into
`int force` (the same class, papered over by truncation) + 2 implicit
double->float narrows in RFV hyperparam assembly -- fixed identically.

NEW seam test C.3d (workspace tests/): drives the PRODUCTION wrapper with
real cfg + real fills -- flag-OFF control, flag-ON pulls, drain-time
per-node fee (exact 2*dfee*1e4 cum_reward shift), Thompson via node_cfg.

Verify: suite 3756/0 (3752 + 4 seam) . determinism GREEN . calls_graph
CLEAN . conformance clean . build green WITH the guard.

Refs: D-428 . reports/2026-08-20-ml-verification-program/ (bandit H-1,
backtest Defects 1-3, plan verdict A-1/A-2/A-3/A-5/A-9/A-10/A-11).

+++ Backtest/BacktestEngine.hpp
@@ -1403,8 +1403,8 @@ static inline void Backtest_RunFullValidation(FullValidationResults *out,
         // subset). Pull from config_used.
         {
             tt::XGBHyperparams hp = tt::XGBHyperparams_Defaults();
-            hp.subsample        = FPN_ToDouble(data->config_used.xgb_subsample);
-            hp.colsample_bytree = FPN_ToDouble(data->config_used.xgb_colsample_bytree);
+            hp.subsample        = (float)FPN_ToDouble(data->config_used.xgb_subsample);   // explicit narrow: XGBHyperparams stores float
+            hp.colsample_bytree = (float)FPN_ToDouble(data->config_used.xgb_colsample_bytree);
             hp.min_child_weight = data->config_used.xgb_min_child_weight;
             hp.seed             = data->config_used.xgb_seed;
             args.snap_max_depth        = hp.max_depth;
